Add generic equalityLengthCheck, moduloLengthCheck methods from BLS utils to precompile utils
