+++ packages/evm/src/precompiles/0d-bls12-g1msm.ts
@@ -3,12 +3,8 @@ import { bytesToHex } from '@ethereumjs/util'
 import { EvmErrorResult, OOGResult } from '../evm.js'
 import { ERROR, EvmError } from '../exceptions.js'
 
-import {
-  gasCheck,
-  leading16ZeroBytesCheck,
-  moduloLengthCheck,
-  msmGasUsed,
-} from './bls12_381/index.js'
+import { gasCheck, leading16ZeroBytesCheck, msmGasUsed } from './bls12_381/index.js'
+import { moduloLengthCheck } from './util.js'
 
 import type { EVMBLSInterface, ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
